Changed headshot detection using `HitboxGroup` instead of `HitboxIndex`
